requirements.txt
python-dotenv
psycopg2-binary
discord.py

